requirements.txt
python-dotenv
pandas
tqdm